+++ SistemaTaller/src/dao/ClienteDAO.java
@@ -1,5 +1,49 @@
 package dao;
 
+import java.util.ArrayList;
+
+import model.Cliente;
+
 public class ClienteDAO {
-    
+    public void insertarCliente(Cliente cliente) {
+        // INSERT SQL 
+    }
+
+    public void eliminarCliente( Cliente cliente) {
+        // DELETE SQL
+    }
+
+    public ArrayList<Cliente> listarClientes() {
+        // SELECT SQL
+        return null;
+    }
+
+    public Cliente getClienteDni() {
+        // SELECT SQL
+        return null;
+    }
+
+    public void actualizarNombreCliente(String nombre) {
+        // UPDATE SQL
+    }
+
+    public void actualizarApellidoCliente(String apellido) {
+        // UPDATE SQL
+    }
+
+    public void actualizarTlfCliente(int apellido) {
+        // UPDATE SQL
+    }
+
+    public void actualizarDireccionCliente(String direccion) {
+        // UPDATE SQL
+    }
+
+    public void actualizarEmailCliente(String email) {
+        // UPDATE SQL
+    }
+
+    public void actualizarDniCliente(String dni) {
+        // UPDATE SQL
+    }
 }
