+++ SistemaTaller/src/dao/EmpleadoDAO.java
@@ -1,5 +1,41 @@
 package dao;
 
+import java.util.ArrayList;
+
+import model.Empleado;
+
 public class EmpleadoDAO {
-    
+    public void insertarEmpleado(Empleado empleado) {
+        // INSERT SQL
+    }
+
+    public void eliminarEmpleado(Empleado empleado) {
+        // DELETE SQL
+    }
+
+    public Empleado getEmpleadoID() {
+        // SELECT SQL
+        return null;
+    }
+
+    public ArrayList<Empleado> listarEmpleados() {
+        // SELECT SQL
+        return null;
+    }
+
+    public void modificarNombreEmpleado(String nombre) {
+        // UPDATE SQL
+    }
+
+    public void modificarApellidoEmpleado(String apellido) {
+        // UPDATE SQL
+    }
+
+    public void modificarTlfEmpleado(int telefono) {
+        // UPDATE SQL
+    }
+
+    public void moficiarIDEmpleado(int id) {
+        // UPDATE SQL
+    }
 }
